Allow bot to remove a friend
From time to time, an organization may wish to remove a friend from the timeline listing. Prior to this PR, the only way to accomplish this was to log into the bot account and manually audit the friend list. This adds the `hubot untappd remove <username>` command to accomplish the same action while using the bot.